Fixes issue 41 Fixes issue 42
And update to the loadQueue not only checks empty, but isDownload() as well. If there are only added titles they are replaced by a fresh retrieve.
An update to getEtag updates Total discs too. This can be used by add and move operations to reference bottom of queue.
Portrait and landscape layouts got a third radial.